Added the functionality surrounding the insertion of players into the program and the sorting of these players into matches and teams
Added the MatchListManager, PlayerQueueManager, and MatchList classes. Also made heavy modifications to the Team and Match classes to accommodate the needs of the added classes. Tests for these classes were added to the main file to test the added classes, but these tests will likely be removed when merging to the main branch.